+++ src/System.IO/ref/System.IO.csproj
@@ -17,6 +17,8 @@
   <ItemGroup>
     <Compile Include="System.IO.cs" />
     <Compile Include="System.IO.Manual.cs" />
+    <!-- ToDo: Remove this P2P reference once new packages are produced and updated -->
+    <ProjectReference Include="..\..\System.Runtime\ref\System.Runtime.csproj" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
