+++ src/System.IO/ref/project.json
@@ -1,8 +1,7 @@
 {
   "dependencies": {
     "System.Runtime": "4.0.20",
-    "System.Text.Encoding": "4.0.0",
-    "System.Threading.Tasks": "4.0.0"
+    "System.Text.Encoding": "4.0.0"
   },
   "frameworks": {
     "netstandard1.7": {
