+++ src/System.Reflection/ref/System.Reflection.csproj
@@ -10,6 +10,8 @@
   <ItemGroup>
     <Compile Include="System.Reflection.cs" />
     <Compile Include="System.Reflection.Manual.cs" />
+    <!-- ToDo: Remove this P2P reference once new packages are produced and updated -->
+    <ProjectReference Include="..\..\System.Runtime\ref\System.Runtime.csproj" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
