+++ src/System.Runtime/ref/System.Runtime.Manual.cs
@@ -44,3 +44,82 @@ namespace System
         public abstract string Name { get; }
     }
 }
+namespace System.Reflection
+{
+    public partial class TypeInfo
+    {
+        // These members are promoted from Type.
+        public abstract System.Reflection.TypeAttributes Attributes { get; }
+        public abstract int GetArrayRank();
+        public abstract System.Type GetElementType();
+        public abstract Type[] GetGenericParameterConstraints();
+        public virtual bool IsSubclassOf(System.Type c) { return default(bool); }
+        public virtual bool IsEquivalentTo(Type other) { return default(bool); }
+        public abstract Type[] GenericTypeArguments { get; }
+        public abstract Type GetGenericTypeDefinition();
+        public abstract Assembly Assembly { get; }
+        public abstract Type BaseType { get; }
+        public abstract bool ContainsGenericParameters { get; }
+
+        public abstract MethodBase DeclaringMethod { get; }
+        public abstract string FullName { get; }
+        public abstract GenericParameterAttributes GenericParameterAttributes { get; }
+        public abstract int GenericParameterPosition { get; }
+        public abstract Guid GUID { get; }
+        public bool HasElementType { get { return default(bool); } }
+        public bool IsAbstract { get { return default(bool); } }
+        public bool IsAnsiClass { get { return default(bool); } }
+        public bool IsArray { get { return default(bool); } }
+        public bool IsAutoClass { get { return default(bool); } }
+        public bool IsAutoLayout { get { return default(bool); } }
+        public bool IsByRef { get { return default(bool); } }
+        public bool IsClass { get { return default(bool); } }
+        public virtual bool IsCOMObject { get { return default(bool); } }
+        public abstract bool IsEnum { get; }
+        public bool IsExplicitLayout { get { return default(bool); } }
+        public abstract bool IsGenericParameter { get; }
+        public abstract bool IsGenericType { get; }
+        public abstract bool IsGenericTypeDefinition { get; }
+        public bool IsImport { get { return default(bool); } }
+        public bool IsInterface { get { return default(bool); } }
+        public bool IsLayoutSequential { get { return default(bool); } }
+        public bool IsMarshalByRef { get { return default(bool); } }
+        public bool IsNested { get { return default(bool); } }
+        public bool IsNestedAssembly { get { return default(bool); } }
+        public bool IsNestedFamANDAssem { get { return default(bool); } }
+        public bool IsNestedFamily { get { return default(bool); } }
+        public bool IsNestedFamORAssem { get { return default(bool); } }
+        public bool IsNestedPrivate { get { return default(bool); } }
+        public bool IsNestedPublic { get { return default(bool); } }
+        public bool IsNotPublic { get { return default(bool); } }
+        public bool IsPointer { get { return default(bool); } }
+        public virtual bool IsPrimitive { get { return default(bool); } }
+        public bool IsPublic { get { return default(bool); } }
+        public bool IsSealed { get { return default(bool); } }
+        public bool IsVisible { get { return default(bool); } }
+        public abstract bool IsSerializable { get; }
+        public bool IsSpecialName { get { return default(bool); } }
+        public bool IsUnicodeClass { get { return default(bool); } }
+        public virtual bool IsValueType { get { return default(bool); } }
+        public abstract string Namespace { get; }
+        public abstract string AssemblyQualifiedName { get; }
+
+        public abstract Type MakeArrayType();
+        public abstract Type MakeArrayType(int rank);
+        public abstract Type MakeByRefType();
+        public abstract Type MakeGenericType(params System.Type[] typeArguments);
+        public abstract Type MakePointerType();
+    }
+
+    // These should be made public when reflection extensibility via contracts is supported on all platforms.
+    // In the meantime, these will be exposed via wrapper factory methods in System.Private.Reflection.Extensibility.
+    public partial struct CustomAttributeNamedArgument
+    {
+        internal CustomAttributeNamedArgument(Type attributeType, string memberName, bool isField, CustomAttributeTypedArgument typedValue) { }
+    }
+
+    public partial struct CustomAttributeTypedArgument
+    {
+        internal CustomAttributeTypedArgument(Type argumentType, object value) { }
+    }
+}
