+++ src/System.Threading.Tasks/ref/System.Threading.Tasks.csproj
@@ -9,6 +9,9 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="System.Threading.Tasks.cs" />
+    <Compile Include="System.Threading.Tasks.Forwards.cs" />
+    <!-- ToDo: Remove this P2P reference once new packages are produced and updated -->
+    <ProjectReference Include="..\..\System.Runtime\ref\System.Runtime.csproj" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
